Adding general projectile controller class
This class takes care of general projectiles fired from turrets. It has basic collision logic that instantiates an effect whenever it hits another collider.

If the projectile detects that it hits a drone, then it will attempt to capture its script, and inflict damage by calling the takedamage function on that drone.